created different files in hook and utils folders
