+++ ecorise-react-app/src/App.js
@@ -1,14 +1,13 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import Dashboard from './Dashboard';
-import LoginScreen from './Login';
+
 
 
 function App() {
   return (
     <Routes>
-      <Route path="/dashboard" element={<Dashboard />} />
-      <Route path="/login" element={<LoginScreen />} />
+      <Route path="/" element={<Dashboard />} />
     </Routes>
   );
 }
