--- ecorise-react-app/src/Login/style.css
@@ -1,159 +1,0 @@
-.login-container {
-  display: flex;
-  width: 100vw;
-  height: 100vh;
-  min-height: 600px;
-  font-family:  Arial;
-}
-
-.login-image {
-  flex: 1 0%;
-  background-image:
-    url(https://muwo1.unibo.it/steamgreenuniboit/wp-content/uploads/sites/6/2021/11/alejo-reinoso-IlmDnJg5cg-unsplash-768x1024.jpg); 
-  background-size: cover;
-  background-position: center;
-}
-
-.input-icon-wrapper {
-  position: relative;
-  display: flex;
-  align-items: center;
-}
-
-.input-icon {
-  position: absolute;
-  left: 474px;    
-  color: #c4c1c1;   
-  pointer-events: none;
-  font-size: 1.1em;
-  z-index: 2;
-  transform: translateY(-50);
-}
-
-.login-input {
-  padding-left: 8px; 
-  width: 150%;
-  box-sizing: border-box;
-}
-
-.toggle-icon {
-  position: absolute;
-  right: 12px;
-  left: auto;
-  z-index: 2;
-  pointer-events: auto;
-  background: none;
-  border: none;
-  font-size: 1.1em;
-}
-.login-input {
-  padding-left: 95px; 
-  width: 100%;
-  box-sizing: border-box;
-}
- 
-.input-icon-wrapper input[placeholder="username"] {
-left: 35px; 
-position: relative;
-}
-.input-icon-wrapper .toggle-icon {
-  position: absolute;
-  right: 10px;
-  left: auto;
-}
-.login-form-section {
-  flex: 1 0%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: #fff;
-}
-
-.login-form {
-  width: 100%;
-  max-width: 500px;
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-}
-
-.login-title {
-  text-align: center;
-  font-size: 2.3rem;
-  color: #8A2324;
-  font-weight: 700;
-  margin-bottom: 42px;
-  margin-top: 1px;
-  letter-spacing: 0.5px;
-}
-
-.login-label {
-  font-size: 1.1rem;
-  color: #353535;
-  font-weight: 500;
-  margin-bottom: 6px;
-  margin-top: 2px;
-}
-
-.login-input {
-  padding: 12px 10px;
-  border: 2px solid #d39b6a;
-  border-radius: 2px;
-  font-size: 1rem;
-  background: #fff;
-  color: #222;
-  margin-bottom: 14px;
-  position: relative;
-
-}
-
-.login-button {
-  width: 100%;
-  background: #8A2324;
-  color: #fff;
-  font-size: 1.1rem;
-  font-weight: 500;
-  padding: 12px 0;
-  border: none;
-  border-radius: 8px;
-  margin-top: 18px;
-  margin-bottom: 16px;
-  cursor: pointer;
-}
-
-.login-button:hover {
-  background: #a03a33;
-}
-
-.login-links {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 18px;
-}
-
-.forgot-password {
-  color: #c0702b;
-  text-decoration: none;
-  font-size: 1rem;
-  margin-bottom: 6px;
-}
-
-.signup-question {
-  color: #222;
-  font-size: 1.08rem;
-}
-
-.signup-link {
-  color: #8A2324;
-  font-weight: 700;
-  text-decoration: none;
-  margin-left: 2px;
-  transition: text-decoration 0.2s;
-}
-
-.signup-link:hover {
-  text-decoration: underline;
-}
-
-
